Create Sensor_simulation_2.c
Q-2 C program in an online compiler,  
1. Use timer to simulate data generated by external sensor. Setup the timer to trigger every 
second and generates random number (0 to 5) of random bytes and adds this to a globally 
accessible data structure.  
2. Separately wake up periodically (every 10s), checks if 50 bytes are stored in the globally 
accessible data structure and prints only the latest 50 bytes (in hex value) and deletes the 
printed bytes from the data structure. 
3. For example, 1st second 4 bytes are added, 2nd second 3 bytes are added and 10th 
second there are 39 bytes are in the buffer. Thus at 10th second data is not printed. At 20th 
second, if there are more than 50 bytes in the buffer, the main thread only prints the latest 
50 bytes and deletes them.